chore: prepare start and endRetrospective for recurrence
Only refactored part of the functionality into different functions.
